refactor(log_viewer): split into 2 focused submodules
- Extract view to view.rs (886 lines)
- Extract tests to tests.rs (586 lines)
- Original mod.rs was 1,475 lines
